requirements.txt
alembic==1.13.1
async-timeout==4.0.3
asyncpg==0.29.0
databases==0.9.0
greenlet==3.0.3
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pyaes==1.6.1
Pyrogram==2.0.106
pyrogram-patch @ git+https://github.com/kotttee/pyrogram_patch.git@ec0a2f4317e19e9f81339a2190eb47c86edfe924
PySocks==1.7.1
python-dotenv==1.0.1
redis==4.5.5
SQLAlchemy==2.0.30
typing_extensions==4.12.2
